Much needed refactor
- introduced HallwayService which handles all business related to creating hallways
- made interfaces consistant (no need for explicitly having a 'public' access modifier
- renamed Uncover Doors to HandleDoorTiles
