+++ src/PDGToolkitCore/Application/IRoomService.cs
@@ -5,7 +5,7 @@ namespace PDGToolkitCore.Application
 {
     internal interface IRoomService
     {
-        public Room MergeRooms(Room r1, Room r2);
+        Room MergeRooms(Room r1, Room r2);
 
         IEnumerable<Room> GetRoomsByPosition(IEnumerable<Room> rooms, Position position);
 
@@ -13,10 +13,6 @@ namespace PDGToolkitCore.Application
 
         IEnumerable<Room> MergeAllRooms(IEnumerable<Room> rooms);
         
-        IEnumerable<Room> CreateDoors(IEnumerable<Room> rooms);
-        
         IEnumerable<Room> TrimSpilledRooms(IEnumerable<Room> rooms);
-
-        public IEnumerable<Tile> UncoverDoorTiles(IEnumerable<Tile> tiles);
     }
 }
