+++ src/PDGToolkitCore/Application/PathFinding/AStarPathFindingService.cs
@@ -13,54 +13,14 @@ namespace PDGToolkitCore.Application.PathFinding
      */
     public class AStarPathFindingService : IPathFindingService
     {
-        private readonly Random random;
-        private Boundary endOfDungeonBoundary;
+        private readonly Boundary endOfDungeonBoundary;
 
-        public AStarPathFindingService(Random random, Settings settings)
+        public AStarPathFindingService(Settings settings)
         {
-            this.random = random;
-
             var dungeonWidth = settings.GridSettings.Width / settings.TileSettings.Size;
             var dungeonHeight = settings.GridSettings.Height / settings.TileSettings.Size;
             const int wallThickness = 1;
             endOfDungeonBoundary = new Boundary(0, dungeonWidth - wallThickness, 0, dungeonHeight - wallThickness);
-            
-        }
-
-        public IEnumerable<Tile> ConstructAllPaths(IEnumerable<Room> rooms)
-        {
-            var allRooms = rooms.ToList();
-            var allTiles = allRooms.SelectMany(r => r.Tiles).ToList();
-            var result = new List<Tile>();
-           
-            var depletedRooms = new List<Room>();
-
-            foreach (var room in allRooms)
-            {
-                depletedRooms.Add(room);
-                var availableRooms = allRooms.FindAll(r => !depletedRooms.Contains(r));
-                if (availableRooms.Any())
-                {
-                    var index = random.Next(availableRooms.Count);
-                    var roomToConnectTo = availableRooms.ElementAt(index);
-
-                    var start = GetDoor(room);
-                    var finish = GetDoor(roomToConnectTo);
-                    result.AddRange(FindPath(start, finish, allTiles));
-                }
-            }
-            
-            allTiles.AddRange(result);
-            var walls = new List<Tile>();
-            
-            foreach (var tile in result)
-            {
-                walls.AddRange(TrySurroundTileWithWallTiles(tile, allTiles));
-            }
-
-            walls = walls.Distinct().ToList();
-            
-            return result.Concat(walls);
         }
 
         public IEnumerable<Tile> FindPath(Tile startTile, Tile finishTile, IEnumerable<Tile> tiles)
@@ -118,26 +78,7 @@ namespace PDGToolkitCore.Application.PathFinding
             Console.Out.WriteLine($"[!] Couldn't find a path!");
             return new List<Tile>();
         }
-
-        private IEnumerable<Tile> TrySurroundTileWithWallTiles(Tile tile, IEnumerable<Tile> tiles)
-        {
-            var potentialSurroundingTiles = new List<Tile>()
-            {
-                new Tile(TileType.Wall, new Position(tile.Position.X, tile.Position.Y - 1)),
-                new Tile(TileType.Wall, new Position(tile.Position.X, tile.Position.Y + 1)),
-                new Tile(TileType.Wall, new Position(tile.Position.X - 1, tile.Position.Y)),
-                new Tile(TileType.Wall, new Position(tile.Position.X + 1, tile.Position.Y)),
-                
-                new Tile(TileType.Wall, new Position(tile.Position.X - 1, tile.Position.Y - 1)),
-                new Tile(TileType.Wall, new Position(tile.Position.X - 1, tile.Position.Y + 1)),
-                new Tile(TileType.Wall, new Position(tile.Position.X + 1, tile.Position.Y - 1)),
-                new Tile(TileType.Wall, new Position(tile.Position.X + 1, tile.Position.Y + 1)),
-            };
-
-            var allPositions = tiles.Select(t => t.Position);
-
-            return potentialSurroundingTiles.Where(t => !allPositions.Contains(t.Position));
-        }
+        
         private List<WeightedTile> GetWalkableTiles(WeightedTile currentTile, WeightedTile targetTile, List<Tile> allTiles)
         {
             var possibleTiles = new List<WeightedTile>
@@ -160,11 +101,6 @@ namespace PDGToolkitCore.Application.PathFinding
                         weightedTile.Position.Y > endOfDungeonBoundary.MinY &&
                         (tileAtPos.IsEmpty() || tileAtPos.IsWalkable());
             }).ToList();
-        }
-        
-        private Tile GetDoor(Room room)
-        {
-            return room.Tiles.First(t => t.Type.Equals(TileType.Door));
         }
     }
 }
