+++ src/PDGToolkitCore/Application/PathFinding/IPathFindingService.cs
@@ -6,7 +6,5 @@ namespace PDGToolkitCore.Application.PathFinding
     public interface IPathFindingService
     {
         IEnumerable<Tile> FindPath(Tile start, Tile finish, IEnumerable<Tile> tiles);
-        
-        IEnumerable<Tile> ConstructAllPaths(IEnumerable<Room> rooms);
     }
 }
