+++ src/PDGToolkitCore/Application/RoomService.cs
@@ -10,18 +10,14 @@ namespace PDGToolkitCore.Application
 {
     internal class RoomService : IRoomService
     {
-        private readonly Random random;
         private readonly Boundary outsideDungeonBoundary;
-        private readonly Boundary doorPositionBoundary;
-        
-        public RoomService(Random random, Settings settings)
+
+        public RoomService(Settings settings)
         {
-            this.random = random;
             var dungeonWidth = settings.GridSettings.Width / settings.TileSettings.Size;
             var dungeonHeight = settings.GridSettings.Height / settings.TileSettings.Size;
             const int wallThickness = 1;
             outsideDungeonBoundary = new Boundary(0, dungeonWidth - wallThickness, 0, dungeonHeight - wallThickness);
-            doorPositionBoundary = new Boundary(1, dungeonWidth - 2 * wallThickness, 1 , dungeonHeight - 2 * wallThickness);
         }
 
         public bool AreRoomsOverlapping(Room firstRoom, Room secondRoom)
@@ -110,19 +106,6 @@ namespace PDGToolkitCore.Application
             return result;
         }
 
-        public IEnumerable<Room> CreateDoors(IEnumerable<Room> rooms)
-        {
-            var allRooms = rooms.ToList();
-            
-            foreach (var room in allRooms)
-            {
-                var allTiles = allRooms.SelectMany(r => r.Tiles).ToList();
-                AllocateDoors(room, allTiles);
-            }
-
-            return allRooms;
-        }
-        
         /**
          * Trims tiles that are spilling out of bounds <see cref="outsideDungeonBoundary"/>
          *  in all <param name="rooms">rooms</param>.
@@ -145,54 +128,6 @@ namespace PDGToolkitCore.Application
             return allRooms;
         }
 
-        private Room AllocateDoors(Room room, List<Tile> allTiles)
-        {
-            var numberOfDoorsForThisRoom = room.NumContainedRooms / 2 + 1;
-            
-            var allWallTiles  = room.Tiles.FindAll(t => t.Type.Equals(TileType.Wall)).ToList();
-            var allDuplicateWallTiles = room.Tiles.GroupBy(x => x)
-                .Where(g => g.Count() > 1)
-                .Select(y => y.Key)
-                .ToList();
-            var uniqueWallTiles = allWallTiles.Except(allDuplicateWallTiles).ToList();
-
-            var allRoomTilesWithoutUniqueWalls = room.Tiles.Except(uniqueWallTiles).ToList();
-
-            var wallTilesWithDoors = ReplaceWallsWithDoorsWherePossible(numberOfDoorsForThisRoom, uniqueWallTiles, allTiles);
-            
-            var allRoomTiles = allRoomTilesWithoutUniqueWalls.Concat(wallTilesWithDoors);
-            
-            room.Tiles.Clear();
-            room.Tiles.AddRange(allRoomTiles);
-            
-            return room;
-        }
-
-        private List<Tile> ReplaceWallsWithDoorsWherePossible(int numberOfDoorsToPlace, List<Tile> uniqueWallTiles, List<Tile> allTiles)
-        {
-            while (numberOfDoorsToPlace != 0)
-            {
-                var index = random.Next(uniqueWallTiles.Count);
-                var randomlyChosenWallTile = uniqueWallTiles.ElementAt(index);
-
-                if (IsTileAValidDoorLocation(randomlyChosenWallTile, allTiles))
-                {
-                    uniqueWallTiles.Add(new Tile(TileType.Door, new Position(uniqueWallTiles.ElementAt(index).Position.X, uniqueWallTiles.ElementAt(index).Position.Y)));
-                    uniqueWallTiles.RemoveAt(index);
-                    numberOfDoorsToPlace--;
-                }
-            }
-
-            return uniqueWallTiles;
-        }
-
-        private bool IsTileAValidDoorLocation(Tile tile, IEnumerable<Tile> tiles)
-        {
-            var allTiles = tiles.ToList();
-            return tile.HasMaxTwoAdjacentWallTiles(allTiles) && !tile.IsOutsideBounds(doorPositionBoundary.MinX, doorPositionBoundary.MaxX,
-                doorPositionBoundary.MinY, doorPositionBoundary.MaxY);;
-        }
-
         /**
          * Removes any tiles that are in the same position as floor tiles.
          */
@@ -206,23 +141,6 @@ namespace PDGToolkitCore.Application
                 tiles.ReplaceTilesWithOtherTile(new Tile(TileType.Floor, tile.Position));
             }
             return tiles;
-        }
-        
-        public IEnumerable<Tile> UncoverDoorTiles(IEnumerable<Tile> tiles)
-        {
-            var allTiles = tiles.ToList();
-            var allDoorTiles = allTiles.FindAll(t => t.Type.Equals(TileType.Door));
-            foreach (var doorTile in allDoorTiles)
-            {
-                if (doorTile.HasTwoAdjacentFloorTiles(allTiles))
-                {
-                    allTiles.ReplaceTilesWithOtherTile(doorTile);
-                    continue;
-                } 
-                allTiles.ReplaceTilesWithOtherTile(new Tile(TileType.Wall, doorTile.Position));
-            }
-
-            return allTiles;
         }
         
         private Dictionary<Position, int> GetPositionsOfDuplicateTiles(List<Tile> tiles)
